Update gcc to version 13
This is a squashed version of #38351

There were a number of places that triggered a dangling-reference warning,
but a lot look like false positives to me..

Also several places had unnecessary std::move() on a string argument in a
constructor that was being passed by reference.

Fix cppcheck warnings

I fixed two warnings and deleted the suppressions, but the unusedScopedObject
warning wasn't obvious how to fix.

Fix maybe-uninitialized warnings from gcc 13

Disable warnings from gcc 13

In gcc 12, any use of _ in Google Mock would generate a
maybe-uninitialized warning, but now it generates an
uninitialized warning, so I've changed over the cmake that
disables those warnings.