Matplotlib 3.8 update  (#38077)
* Due to the ContourSet change in the 3.8 release, we can now check for contour plots by looking for the ContourSet collection in ax.collections.

Co-authored-by: Gui Pereira <gui.maciel-pereira@stfc.ac.uk>
Co-authored-by: Sam Jones <samuel.jones@stfc.ac.uk>

https://matplotlib.org/stable/api/prev_api_changes/api_changes_3.8.0.html#contourset-is-now-a-single-collection

* plotfunctionsTest.py Pass again. Getting the current colour from the axes using Prop_cycler is removed from the newer versions of Matplotlib.

* Back to 3.8

* Asserts possibly no longer necessary. They might be from a time when we had different versions of mpl on different OS.

* Fix test using QuadContourSet

QuadContourSet is now a single collection instead of an artist, so I removed the call to .collections. https://matplotlib.org/3.8.4/api/prev_api_changes/api_changes_3.8.0.html#contourset-is-now-a-single-collection

Co-authored-by: Gui Pereira <gui.maciel-pereira@stfc.ac.uk>

* Fixed 2 failing tests || UnboundLocalError

Turned `current_ax_colour` into a global variable to avoid the error.

* Release notes

* Axes3D.set_frame_on and Axes3D.get_frame_on removed

In Matplotlib V3.8 the `Axes3D.set_frame_on` has no effect on 3D axes and is being removed in favor of `Axes3D.set_axis_on` and `Axes3D.set_axis_off`.

* ContourSet, edgecolor changes from list to array

In the figuremanager.py the PathCollection no longer returns the paths. Instead, it has been moved to a new location where we get a QuadContourSet and change the edge colours that way instead.

In the toolbar.py the edgecolour now receives an array of identical lists instead of a single list, so the first element is picked.

Co-authored-by: Jonathan Haigh <jonathan.haigh@stfc.ac.uk>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Set_Frame_On still necessary for 2D axes.

Added the set_frame_on back to the 2D axes.

* Update docs/source/release/v6.12.0/Framework/Dependencies/New_features/38077.rst

Additional info on release notes

Co-authored-by: James Clarke <139879523+jclarkeSTFC@users.noreply.github.com>

* Update qt/applications/workbench/workbench/plotting/toolbar.py

Removed redundunt line

Co-authored-by: James Clarke <139879523+jclarkeSTFC@users.noreply.github.com>

* Minor change in axisOn  property

Co-authored-by: James Clarke <139879523+jclarkeSTFC@users.noreply.github.com>

* Removed CTestCostData.txt, as it was empty

---------

Co-authored-by: Gui Pereira <gui.maciel-pereira@stfc.ac.uk>
Co-authored-by: Jonathan Haigh <jonathan.haigh@stfc.ac.uk>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: James Clarke <139879523+jclarkeSTFC@users.noreply.github.com>