Handle neutron data for atoms where one isotope = mixture
In cases such as F, one isotope is so dominant that essentially we
would always use the properties of the standard mixture. Oddly, in
this case Mantid will not find neutron data if the atomic mass was
provided.

i.e. Atom('F').neutron() works, but Atom('F', 19).neutron() returns a
dict full of NaN, despite them having exactly the same mass.