Update cppcheck to 2.16.0
This is a squashed version of #38459

fix typo to trigger cppcheck run

Remove cppcheck suppressions temporarily for cppcheck update

Update cppcheck to 2.16.0

Suppress class_X_Y cppcheck warnings.

Restrict cppcheck to unix only. There is no 2.16.0 for Windows.

We can re-add it for Windows in the future if it is reinstated.
For the time being, since nobody is running it locally, we don't really
need it to be installed on Windows.

Sort cppcheck supressions by file name and line number

Group internal cppcheck errors to extract them from the bounty list

Update cppcheck suppressions for v2.16.0