Switch to available constructor in Poco 1.14
This is a squash of #38461

Remove 1.14 Poco pin

Remove old Poco CMake code

These bits of CMake were preventing the solution compiling
on Windows. 1.4.6 is around 10 years old.

Remove old Poco pin

The latest version is 1.14, so no need to keep bit of the pin for
1.12.1

Update pin and comment

The new HTTPSClientSession API won't work with Poco < 1.14, so we
need to pin it to at least 1.14