Set ws to non-distribution after flood correction
The flood correction performs a divide which can result in the corrected
workspace being set as a distribution. This affects the result of the integration
that takes place when normalising by monitors during the
transmission workspace creation. The flood corrected workspace should
not be marked as a distribution, according to the definition that the
integration algorithm is using for this, so we need to set the flag
back to False.