Add algorithm for creating POLREF transmission workspaces (#38466)
* Add new alg ReflectometryISISCreateTransmission

* Add information logging

* Add unit tests for ReflectometryISISCreateTransmission

* Add algorithm documentation

* Add error handling for background subtraction alg

The background subtraction algorithm has a ProcessingInstructions
property, which is the same name as the property on
ReflectometryISISCreateTransmission. This means that error messages can be
confusing if we don't explicitly mention that they have come from the
background subtraction algorithm.

* Change input parameter to take a list of run nums

Use the LoadAndMerge algorithm to load and sum all the runs into a
single workspace as part of the first step.

* Add system test for ReflectometryISISCreateTransmission

* Set ws to non-distribution after flood correction

The flood correction performs a divide which can result in the corrected
workspace being set as a distribution. This affects the result of the integration
that takes place when normalising by monitors during the
transmission workspace creation. The flood corrected workspace should
not be marked as a distribution, according to the definition that the
integration algorithm is using for this, so we need to set the flag
back to False.

* Add release note for PR 38466

* Add issue number to comment

Add link in the comment explaining the workaround for getting workspace group
outputs from child algorithms to the issue that has been opened for this
on GitHub.