Fix test for d-spacing tolerance - was not testing actual behaviour
To do this offset the simulated peaks in one of the two pixels containing the peak - when the center is free to vary (unconstrained by the DIFC ratio or user imposed limits) in the final/second fit, the same solution will be reached.
Hence the I/sig values unchanged, with one exception: I have updated the I/sig test value for Poisson fit - the final fit is still good, just slightly different.