Include AlgoTimeRegister on all operating systems
By only including it on Linux, we need to either edit the docs to only
have examples that only work on Linux, or selectively include the
doc page for that class. It doesn't do any harm to use AlgoTimeRegister
on other operating systems, then the tests can be run on all OS's, and
things are a bit neater.