Fixes for AlgoTimeRegister on Windows
A long on Windows x64 is 4 bytes, but on Linux it is 8 bytes. To hold
nanoseconds we need 8 bytes. By using size_t we'll get an 8 bytes
integer.

File system handling is different on Windows. So in the test, make the
file in the class set up, close it, then open it to write when
required.