Check workspace is retrivable from ADS
This simple change forces the ADS to check whether the workspace name i.e. load_result["DeadTimeTable"] actually exists.
If it doesn't, it gives an error which is caught by the widget, instead of closing the widget and crashing mantid.

Not an ideal solution, but a quick and simple one.