Increase sleep in test algorithm
This test fails intermittently, but I can't reproduce it, even when
building with the CI cmake configuration and running the test 100
times. My best guess is that when the test asserts that the
algorithm hasn't finished yet, it was so quick that it has
finished, so I've increased the sleep in the test algorithm to
make it take longer.