Fix containerOutOfBounds
Cppcheck complained that polynomialCoefficients is accessed when empty.
I think this change to the code is valie because the size of polynomialCoefficients
should be the same as the size of labels.
Also since the main goal of the loop is to run through the coefficients, it makes the intent clearer.