Cppcheck charlie 36 (#38750)
* Delete suppressions

* Fix cppcheck

Most relevant change:

- Marking virtual functions used in constructors as final.
This is fine to do when the class does not have children or when the
version of the virtual function marked as final is intended as the last
override of that function. This is generally the case with setters,
since they are setting a private member variable and tend not to change
in the inherited children.