Add Error message when workspace name is empty
The problem was that when no workspace was added in the interface, the workspace name is a default empty string.
The validating function checkWorkspaceExists() was not adding an error message on this case,
which meant that when other functions evaluated if any error had happened (by doing errorString.empty()),
the result was assumed valid because the error string was empty.

What I did was simply add an error message when the workspace name is empty.
This fixes other cases where the same crash was happening for the same reason.