Change logic to match Poco::Path::isDirectory
Poco::Path::isDirectory only checks if the Poco::Path object has something
labelled as a filename, not necessarily if it exists. Hence to get the
same functionality we need to check if the path exists or not, and also
if it is a directory or not. If it's an existing file then it's no
good.