Doctests: default parameter for CacheDirectory seems to work!
It has been a bit of a headache getting these to work nicely, but it
may have been something to do with spacing between RST lines.

There seems to be a little API change at work here: the default cache
director is now always the default save directory. Previously, it
seems that command-line runs of Mantid would sometimes use the current
directory from which Mantid was launched.

That may be a headscratcher or even breaking for certain user scripts.
In such workflows it would be highly recommended to set an explicit
directory.