Improve Error Description for GSASII Binaries Search
Improve error description when searching for GSASII binaries if they cannot be found. It is not inherently obvious that the outer most parent path of the GSASII installation is required in the settings panel for the model as only GSASIIscriptable is required to import GSAS in call_G2sc.py. The improved error message tells the user where they should set the path too rather than assuming