Pass platform through to standalone
We need the target platform in the script that creates the dmg for
macOS, since it could be either osx-64 or osx-arm64. Previously it was
always osx-64 so could be hard-coded.

I also added the platform name to the dmg file name to avoid a clash
when making both the osx-64 and osx-arm64 standalones.