38882 gsas2scriptable path search revision (#38976)
* Create limited_rglob Method

Create custom implementation of rglob with the method name limited_rglob to recursively search a directory tree structure for a given filename. limited_rglob allows for the setting of a max depth value to limit how deep the directory tree structure is searched to avoid searching uneccesarily deep which could be time-consuming and recourse intensive depening on the path provided

* Create find_gsasii Method to Search for GSASIIscriptable

Replace previous approach which directly set the import path and attempted to import the module with a more robust and dynamic approach. Replacing the previous implementation with find_gsasii allows for controlled depth resursive searching of sub-directories for GSASIIscriptable and does not rely on hard-coded paths which may change across versions of GSAS-II. Moving the previous approach into find_gsasii improves error handling and error visibility by providing more detailed errors which will help diagnose if GSASIIscriptable has been found and there is an import error or if GSASIIscriptable could not be found.

* Improve Error Description for GSASII Binaries Search

Improve error description when searching for GSASII binaries if they cannot be found. It is not inherently obvious that the outer most parent path of the GSASII installation is required in the settings panel for the model as only GSASIIscriptable is required to import GSAS in call_G2sc.py. The improved error message tells the user where they should set the path too rather than assuming

* Create a test class for GSASS-II search and import method

Create a new test class to check a variety of test cases for limited_rglob and find_gsasii

* Release Notes