Avoid casting negative times to unsigned in -- `ornl-next` (#39165)
Avoid casting negative times to unsigned int (#39126)

* Avoid casting negative times to unsigned int

The entries in time_sec can be negative, and casting a negative double
to a uint64_t results in undefined behaviour. The new version will only
cast a negative number to a uint64_t if the time is before 1970, so
most likely not a problem.

* Added release note

* Add test for negative times

* Keep old behaviour for non-negative relative times

By doing this we decrease the amount of changes to times when loading
files, which will help with reproducibility. Changing it to always do
the new method results in times being different (only by a few
nanoseconds).

* Use a signed int to simplify the solution

Co-authored-by: James Clarke <139879523+jclarkeSTFC@users.noreply.github.com>