Abins number selection improvements (#39170)
* Modify Abins "Atoms" parameter to allow numbers without prefix

Currently users can use "Atoms" parameter to select elements by their
symbols, e.g.

         "C,H,O"

and individual atoms by their numbers using a reserved prefix, e.g.

        "atom_1,atom2"

We are permissive about skipping the underscore, but it's not too
difficult to distinguish digits from letters and also permit bare
indices.

        "1,2"

Here this is implemented and the corresponding function is
unit-tested.

I note that isotopes are handled a bit weirdly, you can't specify
e.g. "C12" but if you specify "C" and have an isotopic mixture it
will be separated regardless and give separate workspaces.

Perhaps that's something to revisit in Abins2 which will replace this
"Atoms" interface with a different selection method.

For now, this is a nice little quality-of-life improvement for users
who need to select a lot of atoms!

* Modify Abins "Atoms" parameter to accept number ranges

An even more convenient syntax for atom number input: a range can be
specified with e.g. "1..3" or "1-3".

This will still produce individual output workspaces for now.

* release note

* If atom index range is high-low, quietly swap them and move on

* Update Abins parameter documentation