Improve numerical stability of PhaseQuadMuon (#39178)
* Improve numerical stability of PhaseQuadMuon

This algorithm was inverting a 2x2 matrix manually, which resulted in
numerically unstable results. On different CPU architectures, the
results are wildly different, and even doing the calculation in a
slightly different order changes the results. By using EigenMatrix, we
get a more stable inverse.

* Added a release note

* Improve logic clarity

Co-authored-by: Adri Diaz <146007827+adriazalvarez@users.noreply.github.com>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Use constant for comparison delta

* Change docs example to have well-defined solution

This arbitrary example in the docs was not well-defined. The solution
required inverting a matrix that was close to being singular, resulting
in unreliable answers.

---------

Co-authored-by: Adri Diaz <146007827+adriazalvarez@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>