Remove GSL error code workaround
The error GSL_ENOPROG is an error code indicating that the solver has
stopped, but we were trying to force it to carry on. One potential
problem is that if we do that, and the step size is within tolerance,
then later on in the method we will report a success, (incorrectly).

I've also changed the test to handle the case where the error code
could be different, which is what happens on ARM. The solver stops for
a different reason to x86_64.