Fix casting negative double to unsigned int in GenerateGroupingPowder -- `ornl-next` (#39222)
Fix casting negative double to unsigned int in GenerateGroupingPowder (#39205)

* Fix casting negative double to unsigned int

A size_t is an unsigned int, but the way this method was casting
resulted in a negative value being cast to a large positive value. It
now uses an int64_t for the intermediate step instead.

* Add release note

* Results changed in this test

* Avoid int -> double conversion

Without this cast, the second component of the sum will be converted
from an int64_t to a double, resulting in a compiler warning and
potential loss of precision.

Co-authored-by: James Clarke <139879523+jclarkeSTFC@users.noreply.github.com>