handle new PRECALL instruction in funcinspect
For some reason OSIRISDiffactionReductionTest was failing when running all of it's unit test.
The first three would pass, then the rest fail (no matter which you commented out etc).
Running them individually was fine.
I still don't understand why this was happening, but it was caused by line 374 of OSIRISDiffactionReduction
where some workspaces are summed. This calls the op_wrapper in _workspaceops which calls lhs_info()

Through debugging I worked out that when the tests were failing, frame.f_lasti was pointing to PRECALL instead of CALL.
So I've changed process_frame to change last_i to point at CALL when it is at PRECALL.