Change FileLoaderRegistry chooseLoader for HDF5 files to only check LegacyNexus loaders when confidence if too low (#39280)
### Description of work

This initial bug report states that `test = Load("NOM_214940",
MetaDataOnly=True)` is very slow, much slower than in Mantid 6.12,
`NOM_214940` is a 16Gb file.

You can reduce this to
```python
from mantid.simpleapi import *
from mantid.api import FileLoaderRegistry
FileLoaderRegistry.Instance().chooseLoader("/SNS/NOM/IPTS-34487/nexus/NOM_214940.nxs.h5")
```
and see that it is the `chooseLoader` that is slow.

The issue is that creating the `LegacyNexusDescriptor` is very slow for
large files so we only want to do this when needed.

So this change only checks that `LegacyNexus` loaders when needed. That
is when confidence from the `NexusDescriptor` loaders is less than 80,
so not loaded by `LoadEventNexus`, `LoadNexusProcessed`,
`LoadMuonNexusV2` or `LoadMD`.

Unfortunately we can't just check `LegacyNexusDescriptor` when
confidence from `NexusDescriptor` is `0` because for some files (_e.g._
`GPS5397.NXS` and `GPD900.nxs` from the systemtest data) show a
confidence of `75` from `LoadISISNexus` (a `NexusDescriptor` loader) but
actually needs to be loader by `LoadMuonNexus` (a
`LegacyNexusDescriptor` loader).

#### Summary of work
<!-- Please provide a short, high level description of the work that was
done.
-->

<!-- Why has this work been done? If there is no linked issue please
provide appropriate context for this work.
#### Purpose of work
This can be removed if a github issue is referenced below
-->

Fixes
[11112](https://ornlrse.clm.ibmcloud.com/ccm/resource/itemName/com.ibm.team.workitem.WorkItem/11112)

<!-- If the original issue was raised by a user they should be named
here. Do not leak email addresses
**Report to:** [user name]
-->

#### Further detail of work
<!-- Please provide a more detailed description of the work that has
been undertaken.
-->

### To test:

Running this on analysis should be fast again

```python
test = Load("NOM_214940", MetaDataOnly=True)
```

<!-- Instructions for testing.
There should be sufficient instructions for someone unfamiliar with the
application to test - unless a specific
reviewer is requested.
If instructions for replicating the fault are contained in the linked
issue then it is OK to refer back to these.
-->


*This does not require release notes* because **the change to
FileLoaderRegistry that made it slow was only done in this release.**
Change was made in #38786


<!-- Ensure the base of this PR is correct (e.g. release-next or main)
Finally, don't forget to add the appropriate labels, milestones, etc.!
-->

---

### Reviewer

Please comment on the points listed below ([full
description](http://developer.mantidproject.org/ReviewingAPullRequest.html)).
**Your comments will be used as part of the gatekeeper process, so
please comment clearly on what you have checked during your review.** If
changes are made to the PR during the review process then your final
comment will be the most important for gatekeepers. In this comment you
should make it clear why any earlier review is still valid, or confirm
that all requested changes have been addressed.

#### Code Review

- Is the code of an acceptable quality?
- Does the code conform to the [coding
standards](http://developer.mantidproject.org/Standards/)?
- Are the unit tests small and test the class in isolation?
- If there is GUI work does it follow the [GUI
standards](http://developer.mantidproject.org/Standards/GUIStandards.html)?
- If there are changes in the release notes then do they describe the
changes appropriately?
- Do the release notes conform to the [release notes
guide](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)?

#### Functional Tests

- Do changes function as described? Add comments below that describe the
tests performed?
- Do the changes handle unexpected situations, e.g. bad input?
- Has the relevant (user and developer) documentation been
added/updated?

Does everything look good? Mark the review as **Approve**. A member of
`@mantidproject/gatekeepers` will take care of it.

### Gatekeeper

If you need to request changes to a PR then please add a comment and set
the review status to "Request changes". This will stop the PR from
showing up in the list for other gatekeepers.