Add ARM option to building packages on Jenkins (#38990)
This adds the option to build a native package for Arm Macs. It will be
listed as a new option called `osx-arm64`, with the existing option
remaining as `osx-64`.

The `osx-arm64` platform can only run on Arm macOS machines, and it will
compile native Arm code instead of cross-compiling for `x86_64`. The
platform is used for specifying the cmake preset, which is why I added
the `osx-arm64-ci` preset.

Needed for #38973 
Fixes #38995 

In various places the desired platform is passed as an argument. We
cannot simply use the native platform, because any ARM build machine
could be either targeting `osx-64` or `osx-arm64`, so we need to know
what was selected.

I've simplified some code in `mamba-utils`, and deleted the bit that was
forcing `osx-64` on Macs.

When archiving the environment logs we get the platform from the
configuration of the current conda environment. This is because that
script is called from the build script in the conda recipe, which
doesn't know about options specified in Jenkins, so it seemed like the
best way of getting that information.

The `jemalloc` library is pinned to version `5.2.0` for Linux, but
`5.2.1` is the first version with an Arm build. Luckily we don't need
`jemalloc` on macOS anyway.

`quickbayes` is now a conda package, and is a `noarch` package, so this
isn't a problem any more.

Several tests (especially anything that involves fitting) generate
different numbers on ARM machines. In some places I've adjusted the
tolerance, but in others (especially some system tests), I've just
skipped the test on ARM.

### To test:
- On an Arm machine, check that the package runs okay, and is running
natively, not under Rosetta.
- On an Arm machine, check that the old (Intel) package runs okay, and
is running under Rosetta.
- Try running the `Core_Team_test_pipeline` with the `osx-arm64` option
and the `osx-64` option.
- Build and test standalone packages on all platforms.

When testing on an ARM machine, opening Activity Monitor will show if
the process is 'Intel' or 'Apple', i.e. x86 or ARM.