Remove prefix input and automate it from ws group name
Including one argument for prefix and another for the group name did not
make sense because there was nothing to be gained from the group name
being different from the prefixes.
On the other hand, this could easily generate confusion and was prone to
errors where the algorithm was run again with the same prefix but
different group name or vice versa.

Also the decision to store the workspace names and create a group at the
end is a conscious one. Enforcing the creation of a group through the
child algorithm is preferred because it prevents bugs (for example
groups not being correctly overridden). This child algorithm cannot be
initialized without workspaces, hence the decision to use a vector of
strings to store the workspaces to group.