Add new algorithm CreateBootstrapWorkspaces (#39324)
### Description of work
A new simple algorithm for creating batches of bootstrap workspaces.
Requested by Vesuvio scientists, but applicable in general to any
science area.

#### Summary of work
The idea of bootstrap is to simulate sample workspaces by sampling with
replacement from the input workspace. This algorithm introduces two ways
of doing this: the first by sampling from a gaussian distribution at
*each point* (the distribution has the mean at the point and deviation
the error at the point) and the second way by resampling with repetition
the spectra indices. This latter is more particular for the Vesuvio
routines and does not make sense when the rows in the workspace are not
somewhat interchangeable.

This algorithm was requested by the Vesuvio instrument scientists, which
use Bootstrap frequently in some of their analysis.

#### Further detail of work
Most of the code is boilerplate for algorithms. The algorithm itself is
rather straightforward. I would recommend to read the documentation page
first and then to look at my implementation of it. The scientists
provided the original implementation in Python (maybe I should have kept
this algorithm in Python for simplicity sake, I didn't because I noticed
that a very similar algorithm, `CreateMonteCarloWorkspace`, was already
written in C++):