Fix crash in Muon results table (#39338)
### Description of work
This PR fixes a crash when retrieving certain logs form the muon results
table in the Muon Analysis Interface.
Check #39263
#### Summary of work
<!-- Please provide a short, high level description of the work that was
done.
-->
There's a certain set of allowed logs that can go into the results
table. These logs are mostly time series, but can also be strings or
numbers. If these logs are time series, they are averaged. Additionally,
logs that are not time series, are attempted to be converted to float.
If the selected log is not available in a given selected fitted
workspace, there was a recursive call to all the workspaces used in the
fit, until the log was found. At some point the log named `periods`
appears to have been changed for some Muon instruments, and now is in
the form of an integer time series, which can't be detected by this
function, therefore the call to this function gets trapped in an
infinite loop that crashes Mantid.
Two things have been done in this PR: 
- Fix the crash, all the logs that can be converted to a float or
averaged to a float are checked against their mantid types ( ie.
IntTimePropertiySeries, FloatTimePropertySeries, ..) . All the fitted
workspaces are stored in a list and the list is iterated in a for loop
to check against these ws if it is not available in the selected fit ws,
avoiding recursive calls.
- As per required from the scientist that submitted the error report, I
have added the log `current_period` to the set of allowed logs into the
results table.

Fixes #39263.