Add capacity to load polarized files to LoadNXcanSAS (#39331)
### Description of work

Following from #39118, this PR extends the LoadNXcanSAS algorithm to
load data saved with SavePolarizedNXcanSAS algorithm.

#### Summary of work

The main changes are: 

- There's been a substantial refactor of the `LoadNXcanSAS` class to
extend it to load polarized data. The original algorithm differentiated
between extracting 1d and 2d data from the h5 file datasets. In this
case, I have wanted to homogenize the way it is done, so there is only
one function to load every kind of data, and the data extracted is
managed through the offset vector on the hyper slab as well as the
vector that controls the slice shape.
- I needed to reuse some of the data structures defined in the
`NXcanSASHelper` file that were used in save algorithms. In order to
organize better the code, I have renamed the `NXcanSASHelper` file to
`SaveNXcanSASHelper `and move several data structures and functions of
common use to a separate file called `NXcanSASUtil`.
- One noticeable change is that I have generally preferred to make sure
a given dataset or group in the nexus file exists before trying to open
it. This is in contrast to the philosophy that was in the file (access
first, throw if it doesn't exists). This was creating a lot of long and
confusing warnings after throwing in the tests and when loading some
files in the algorithm.
- Every polarization state in the NXcanSAS file is loaded into a
separate workspace, and all are grouped into a group workspace. To
distinguish the polarization state, I have added a sample log in the
workspace with the NXcanSAS polarization label.

Fixes #38526

### To test:

- Using the test code of #39118 . Generate few polarization files. Try
to load the save files using the load dialog of Mantid (it should detect
that the loader algorithm is `LoadNXcanSAS`)
- Generate a non-polarized file and save using `SaveNXcanSAS`, then
check is loaded correctly with `LoadNXcanSAS`:
```

# import mantid algorithms, numpy and matplotlib
from mantid.simpleapi import *
import matplotlib.pyplot as plt
import numpy as np

#import Command Interface from sans library
from sans.command_interface.ISISCommandInterface import *

LOQ()
MaskFile('MaskFile.toml')

AssignSample('LOQ74044.nxs')
TransmissionSample('LOQ74024.nxs', 'LOQ74014.nxs')
AssignCan('LOQ74019.nxs')
TransmissionCan('LOQ74020.nxs', 'LOQ74014.nxs')

#Performs reduction and saves data to path using SaveRKH and SaveNexus algorithms
WavRangeReduction(2.2, 10, output_name='test_load',saveAlgs={'SaveNXCanSAS': 'nxs'})
ws_loaded=LoadNXcanSAS(Filename="test_load_main_1D_2.2_10.0.h5")
```