39139 larger gsasii refactor (#39240)
- Try to encapsulate and centralise logic for managing GSAS-II input
parameters, paths, and configurations into a class (`GSAS2Handler`).
- Encapsulate related attributes into logical groups using dataclasses
to improve readability and reduce the size of the inits for
`GSAS2Handler` and `GSAS2Model`.
- Add additional tests to check GSAS2 is correctly loaded and necessary
parameters are handles correctly before GSAS2 is imported.
- Improve modularity and maintainability of model through type driven
refactoring - (this is not an extensive refactor as this would be out of
scope. Additional refactoring can be done in the future to further
improve modularity and maintainability).