Ensure proton charge returned consistently across datasets with period data (#39322)
#### Summary of work

A bug was introduced in #39029. For any dataset that has period data,
the log `proton_charge_by_period` is used to return the proton charge
for the current period.

When runs are added. the `proton_charge_by_period` log is not summed; An
erroneous proton charge is therefore subsequently returned.

To resolve this, this PR enables the elementwise addition of the
`proton_charge_by_period` log during run addition.

Using a bool property was causing warnings to be raised during workspace
save, as bool properties are not supported. The unfiltered flag has
therefore been changed to an int.

Also, the check that flags for period data filtering has been moved
outside of ` if
(!workspace->run().hasProperty(workspace->run().getProtonChargeLogName()))`
to account for datasets that come pre-loaded with this log (vesuvio).

This PR fixes a number of the more prelevant issues, but leaves some
inconsistencies. An issue to fix these issues in an upcoming has been
raised:
https://github.com/mantidproject/mantid/issues/39336

Some release notes have been added which I owed to other PRs.