Use a Relative Tolerance in SXD System Test (#39394)
The values in the test are pretty large, so a relative tolerance is more
appropriate here anyway. Also it makes the test pass on macOS, which is
holding up the nightly.

The change also makes use of `np.testing.assert_allclose` rather than an
an `assertTrue` (thanks @warunawickramasingha for finding that
function). This returns a really helpful error message so would be good
to use elsewhere when we're making numpy comparisons (particularly for
fitting).