NeXus File uses NexusDescriptor (#39362)
This adds a `NexusDescriptor` object inside of `NeXusFile` to track the
file tree. This means user code can check if a group or dataset exists
in a file without having to touch disk.

#### Purpose of work

It is not going to be possible to finish PR #39213 before code freeze
and release 6.13. This was seen as one marginal improvement that can be
implemented in time for code freeze.

Part of #38332