Remove unused NXAccess and pre-compiler options (#39388)
This removes a variety of unused (by mantid) options and related from
napi and things below it. Specific options removed are:
* `NXACC_NOSTRIP`
* `NXACC_CHECKNAMESYNTAX`
* `NXACC_CREATEXML`
* `HAVE_HDF5` and `WITH_HDF5` because it is hdf5 only
* `napiconfig.h`

Then a first attempt at making `NXopen` clearer since it only has one
filetype to deal with.

Refs #38332

### To test:

Basically a refactor so if all the tests pass it works.

*This does not require release notes* because it changes nothing user
facing.

<!-- Ensure the base of this PR is correct (e.g. release-next or main)
Finally, don't forget to add the appropriate labels, milestones, etc.!
-->