Remove unused napimount and external link methods from `napi` (#39408)
This removes the unused external link behavior that is part of `napi`.
This functionality is never used within Mantid, and so can be safely
dropped.

#### Purpose of work

These large functions and blocks that serve no purpose make it harder to
track the behavior inside `napi` of things like open/close positions.

Part of #38332