Remove unused functions from napi (#39409)
This removes a variety of functions in napi that were not being used by
Nexus::File
* NXlinkexternal and NXlinkexternaldataset
* int32_t variants of NXcompmakedata, NXmalloc, NXgetinfo, NXgetrawinfo,
NXgetslab, NXmakedata
* The ability to set the cache size for the hdf5 file

`NXmalloc` remains because it is used for getting attributes which don't
currently allow `int64_t` dimensions.
`NXmakedata64` is currently the only way to create NXunlimited
dimensions.

Refs #38332 and [EWM
11449](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20%28Change%20Management%29#action=com.ibm.team.workitem.viewWorkItem&id=11449)

### To test:

This removes unused functionality and should not affect any tests.

*This does not require release notes* because it deletes unused c++
code.