Remove the X in NeXusException and NeXusIOHelper (#39418)
To reduce diversity in namespaces and filenames, this renames
`NeXus::Exception` to `Mantid::Nexus::Exception` (new header name is
`Exception.h`) and `Mantid::NeXus::NeXusIOHelper` to
`Mantid::Nexus::IOHelper` (header file unchanged).

Additionally, a lot of exceptions are now caught by const reference.

Refs #38332 and
[EWM11399](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20(Change%20Management)#action=com.ibm.team.workitem.viewWorkItem&id=11399)

### To test:

This is a mechanical rename of class/namespace. No functionality was
changed.

*This does not require release notes* because only names have been
changed (to protect the innocent?).