Change Nexus path to address (#39417)
### Description of work

To help maintain good order and civility, `path` has been replaced with
`address` whenever referring to a location within a Nexus file.

<!-- Please provide a short, high level description of the work that was
done.
-->

<!-- Why has this work been done? If there is no linked issue please
provide appropriate context for this work.
#### Purpose of work
This can be removed if a github issue is referenced below
-->

Done in support of #38332 <!-- and fix #xxxx or close #xxxx xor resolves
#xxxx. One line per issue fixed. -->
<!-- alternative
*There is no associated issue.*
-->

<!-- If the original issue was raised by a user they should be named
here. Do not leak email addresses
**Report to:** [user name]
-->

<!-- Please provide a more detailed description of the work that has
been undertaken.
-->

### To test:

This is a refactor to rename variables and functions. All existing tests
should pass.

*This does not require release notes* because this is an internal change
only.