Use existing euphonic method to iterate over metadata
Spectrum collection store metadata as a mixture of common attributes
and line-specific, e.g.

{"element": "Na",
 "scattering": "Coherent",
 "line_data": [{"atom_index": 1}, {"atom_index": 2}]
}

The constructors and iteration methods on these classes split and
merge as needed. When we want to only iterate over the metadata, it is
wasteful to iterate over the whole collection (i.e. spawning new
Spectrum objects) so we use the iter_metadata() method which doesn't
touch the main data arrays.

This wasn't always available in euphonic so was initially duplicated here.