Rework autoconvolution to use Spectrum1DCollection
This method was working with lists of spectra in order to avoid some
overhead in constructing and iterating over the sequences. Between an
upstream refactor of the .y_data method and the addition of an "usafe"
option to .from_spectra, this should no longer be necessary.

This allows us to work reasonably efficiently without using private
attributes.